pyproject.toml
[tool.poetry]
name = "discord-bot-template"
version = "0.1.0"
description = ""
authors = ["Chris Wood <woodc@us.ibm.com>"]
readme = "README.md"
packages = [{include = "discord_bot_template"}]

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.2.3"
requests = "^2.30.0"


[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
black = "^23.3.0"
pre-commit = "^3.3.2"
pytest = "^7.3.1"
pylint = "^2.17.4"
pylint-pydantic = "^0.1.8"

[tool.bandit]
exclude_dirs = ["tests", "env", "__pycache__", ".pytest_cache"]
skips = ["B104"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
